Make distributables for shipping
