+++ .gitignore
@@ -7,7 +7,7 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 node_modules
-dist
+dist/
 dist-ssr
 *.local
