+++ src/SideBar.jsx
@@ -20,7 +20,7 @@ const SideBar = () => {
                             <li className={isActive ? 'active' : ''}>
                                 <div>
                                 <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <g clip-path="url(#clip0_483_1304)">
+                                    <g clipPath="url(#clip0_483_1304)">
                                     <path d="M8.53905 3.46736C8.49513 3.42563 8.43669 3.40234 8.37591 3.40234C8.31513 3.40234 8.2567 3.42563 8.21277 3.46736L2.78177 8.61979C2.7587 8.64171 2.74035 8.66804 2.72783 8.69719C2.7153 8.72635 2.70886 8.75773 2.7089 8.78942L2.70802 13.6252C2.70802 13.8738 2.80747 14.1123 2.98451 14.2881C3.16155 14.4639 3.40166 14.5627 3.65203 14.5627H6.48701C6.61219 14.5627 6.73225 14.5133 6.82076 14.4254C6.90928 14.3375 6.95901 14.2182 6.95901 14.0939V10.1095C6.95901 10.0474 6.98388 9.98777 7.02813 9.94381C7.07239 9.89986 7.13242 9.87517 7.19501 9.87517H9.55504C9.61763 9.87517 9.67766 9.89986 9.72192 9.94381C9.76618 9.98777 9.79104 10.0474 9.79104 10.1095V14.0939C9.79104 14.2182 9.84077 14.3375 9.92929 14.4254C10.0178 14.5133 10.1379 14.5627 10.263 14.5627H13.0968C13.3472 14.5627 13.5873 14.4639 13.7644 14.2881C13.9414 14.1123 14.0409 13.8738 14.0409 13.6252V8.78942C14.0409 8.75773 14.0345 8.72635 14.0219 8.69719C14.0094 8.66804 13.9911 8.64171 13.968 8.61979L8.53905 3.46736Z" 
                                     fill={isActive ? "#FB991C" : "#FFFFFF"}/>
                                     <path d="M15.3053 7.65381L13.0987 5.55732V2.37598C13.0987 2.25166 13.0489 2.13243 12.9604 2.04452C12.8719 1.95661 12.7518 1.90723 12.6267 1.90723H11.2106C11.0855 1.90723 10.9654 1.95661 10.8769 2.04452C10.7884 2.13243 10.7386 2.25166 10.7386 2.37598V3.31348L9.02999 1.69102C8.87009 1.53047 8.63232 1.43848 8.37537 1.43848C8.11931 1.43848 7.88213 1.53047 7.72224 1.69131L1.44752 7.65322C1.26403 7.829 1.24102 8.11816 1.40799 8.30859C1.44992 8.35666 1.50126 8.39575 1.55888 8.42349C1.61651 8.45122 1.67922 8.46703 1.74318 8.46993C1.80715 8.47284 1.87105 8.46278 1.93098 8.44039C1.99092 8.41799 2.04564 8.38371 2.09181 8.33965L8.21312 2.53066C8.25704 2.48894 8.31548 2.46565 8.37626 2.46565C8.43704 2.46565 8.49548 2.48894 8.5394 2.53066L14.6613 8.33965C14.7515 8.42553 14.8723 8.4724 14.9972 8.46998C15.1221 8.46757 15.241 8.41606 15.3277 8.32676C15.5088 8.14043 15.4938 7.83281 15.3053 7.65381Z"
